v1.0.3: fix rails 6 reloader
