+++ lib/dhs/concerns/autoload_records.rb
@@ -18,6 +18,9 @@ module AutoloadRecords
       end
 
       class Middleware
+
+        MODEL_FILES = 'app/models/**/*.rb'.freeze
+
         def initialize(app)
           @app = app
         end
@@ -27,24 +30,25 @@ module AutoloadRecords
           @app.call(env)
         end
 
-        def self.model_files
-          Dir.glob(Rails.root.join('app', 'models', '**', '*.rb'))
-        end
 
         def self.require_direct_inheritance
-          model_files.sort.map do |file|
-            next unless File.read(file).match('DHS::Record')
-            require_dependency file
-            file.split('models/').last.gsub('.rb', '').classify
-          end.compact
+          Rails.application.reloader.to_prepare do
+            Dir.glob(Rails.root.join(MODEL_FILES)).each do |file|
+              next unless File.read(file).match('DHS::Record')
+              require_dependency file
+              file.split('models/').last.gsub('.rb', '').classify
+            end.compact
+          end
         end
 
         def self.require_inheriting_records(parents)
-          model_files.each do |file|
-            file_content = File.read(file)
-            next if parents.none? { |parent| file_content.match(/\b#{parent}\b/) }
-            next if file_content.match?('extend ActiveSupport::Concern')
-            require_dependency file
+          Rails.application.reloader.to_prepare do
+            Dir.glob(Rails.root.join(MODEL_FILES)).each do |file|
+              file_content = File.read(file)
+              next if parents.none? { |parent| file_content.match(/\b#{parent}\b/) }
+              next if file_content.match?('extend ActiveSupport::Concern')
+              require_dependency file
+            end
           end
         end
