+++ lib/dhs/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module DHS
-  VERSION = '1.0.2'
+  VERSION = '1.0.3'
 end
